+++ tests/test_vmf.py
@@ -122,42 +122,41 @@ def test_by_target() -> None:
     assert vmf.by_target == {'some_name': {ent1}, None: {vmf.spawn}}
 
     ent2 = Entity(vmf, {'classname': 'info_target', 'targetname': 'some_name'})
-    assert vmf.by_target == {'some_name': {ent1}, None: {vmf.spawn}} # Not yet included.
+    assert vmf.by_target == {'some_name': {ent1}, None: {vmf.spawn}}  # Not yet included.
     vmf.add_ent(ent2)
     assert vmf.by_target == {'some_name': {ent1, ent2}, None: {vmf.spawn}}
 
     ent1['targetname'] = 'another'
     assert vmf.by_target == {'some_name': {ent2}, 'another': {ent1}, None: {vmf.spawn}}
     del ent2['targetname']
-    assert vmf.by_target == {'another': {ent1}, None: {vmf.spawn, ent2}, 'some_name': set()}
+    assert vmf.by_target == {'another': {ent1}, None: {vmf.spawn, ent2}}
     ent2['targetname'] = 'some_name'
     ent1.remove()
-    assert vmf.by_target == {'another': set(), 'some_name': {ent2}, None: {vmf.spawn}}
+    assert vmf.by_target == {'some_name': {ent2}, None: {vmf.spawn}}
 
 
 def test_by_class() -> None:
     """Test the behaviour of the by_class lookup mechanism."""
     vmf = VMF()
-    assert vmf.by_class == {'': set(), 'worldspawn': {vmf.spawn}}
+    assert vmf.by_class == {'worldspawn': {vmf.spawn}}
 
     ent1 = vmf.create_ent('info_target')
-    assert vmf.by_class == {'': set(), 'worldspawn': {vmf.spawn}, 'info_target': {ent1}}
-    assert list(vmf.by_class) == IsList('', 'worldspawn', 'info_target', check_order=False)
-    assert list(vmf.by_class.keys()) == IsList('', 'worldspawn', 'info_target', check_order=False)
+    assert vmf.by_class == {'worldspawn': {vmf.spawn}, 'info_target': {ent1}}
+    assert list(vmf.by_class) == IsList('worldspawn', 'info_target', check_order=False)
+    assert list(vmf.by_class.keys()) == IsList('worldspawn', 'info_target', check_order=False)
     assert vmf.by_class['info_target'] == {ent1}
 
     ent2 = Entity(vmf, {'classname': 'info_target'})
-    assert vmf.by_class == {'': set(), 'worldspawn': {vmf.spawn}, 'info_target': {ent1}}  # Not yet included.
+    assert vmf.by_class == {'worldspawn': {vmf.spawn}, 'info_target': {ent1}}  # Not yet included.
     vmf.add_ent(ent2)
-    assert vmf.by_class == {'': set(), 'worldspawn': {vmf.spawn}, 'info_target': {ent1, ent2}}
+    assert vmf.by_class == {'worldspawn': {vmf.spawn}, 'info_target': {ent1, ent2}}
 
     ent1['classname'] = 'math_counter'
-    assert vmf.by_class == {'': set(), 'worldspawn': {vmf.spawn}, 'info_target': {ent2}, 'math_counter': {ent1}}
+    assert vmf.by_class == {'worldspawn': {vmf.spawn}, 'info_target': {ent2}, 'math_counter': {ent1}}
     with pytest.raises(KeyError):  # Not allowed.
         del ent2['classname']
     ent2.remove()
     assert vmf.by_class == {
-        '': set(), 'info_target': set(),
         'worldspawn': {vmf.spawn},
         'math_counter': {ent1},
     }